Layer 133 API

+++ tl/src/main/kotlin/com/github/badoualy/telegram/tl/api/TLAbsChat.kt
@@ -4,15 +4,15 @@ import com.github.badoualy.telegram.tl.core.TLObject
 
 /**
  * Abstraction level for the following constructors:
- * * [channel#cb44b1c][TLChannel]
- * * [channelForbidden#289da732][TLChannelForbidden]
- * * [chat#d91cdd54][TLChat]
- * * [chatEmpty#9ba2d800][TLChatEmpty]
- * * [chatForbidden#7328bdb][TLChatForbidden]
+ * * [channel#8261ac61][TLChannel]
+ * * [channelForbidden#17d493d5][TLChannelForbidden]
+ * * [chat#41cbf256][TLChat]
+ * * [chatEmpty#29562865][TLChatEmpty]
+ * * [chatForbidden#6592a1a7][TLChatForbidden]
  *
  * @author Yannick Badoual yann.badoual@gmail.com
  * @see <a href="http://github.com/badoualy/kotlogram">http://github.com/badoualy/kotlogram</a>
  */
 abstract class TLAbsChat : TLObject() {
-    abstract var id: Int
+    abstract var id: Long
 }
