+++ tl/src/main/kotlin/com/github/badoualy/telegram/tl/api/TLAbsChatParticipants.kt
@@ -4,12 +4,12 @@ import com.github.badoualy.telegram.tl.core.TLObject
 
 /**
  * Abstraction level for the following constructors:
- * * [chatParticipants#3f460fed][TLChatParticipants]
- * * [chatParticipantsForbidden#fc900c2b][TLChatParticipantsForbidden]
+ * * [chatParticipants#3cbc93f8][TLChatParticipants]
+ * * [chatParticipantsForbidden#8763d3e1][TLChatParticipantsForbidden]
  *
  * @author Yannick Badoual yann.badoual@gmail.com
  * @see <a href="http://github.com/badoualy/kotlogram">http://github.com/badoualy/kotlogram</a>
  */
 abstract class TLAbsChatParticipants : TLObject() {
-    abstract var chatId: Int
+    abstract var chatId: Long
 }
