+++ tl/src/main/kotlin/com/github/badoualy/telegram/tl/api/TLChatPhotoEmpty.kt
@@ -20,6 +20,6 @@ class TLChatPhotoEmpty : TLAbsChatPhoto() {
         return true
     }
     companion object  {
-        const val CONSTRUCTOR_ID: Int = 0x37c1011c.toInt()
+        const val CONSTRUCTOR_ID: Int = 0x37c1011c
     }
 }
