+++ tl/src/main/kotlin/com/github/badoualy/telegram/tl/api/TLDocumentAttributeAnimated.kt
@@ -20,6 +20,6 @@ class TLDocumentAttributeAnimated : TLAbsDocumentAttribute() {
         return true
     }
     companion object  {
-        const val CONSTRUCTOR_ID: Int = 0x11b58939.toInt()
+        const val CONSTRUCTOR_ID: Int = 0x11b58939
     }
 }
