+++ tl/src/main/kotlin/com/github/badoualy/telegram/tl/api/TLEncryptedFile.kt
@@ -83,6 +83,6 @@ class TLEncryptedFile() : TLAbsEncryptedFile() {
                 && keyFingerprint == other.keyFingerprint
     }
     companion object  {
-        const val CONSTRUCTOR_ID: Int = 0x4a70994c.toInt()
+        const val CONSTRUCTOR_ID: Int = 0x4a70994c
     }
 }
