+++ tl/src/main/kotlin/com/github/badoualy/telegram/tl/api/TLInputDocumentEmpty.kt
@@ -20,6 +20,6 @@ class TLInputDocumentEmpty : TLAbsInputDocument() {
         return true
     }
     companion object  {
-        const val CONSTRUCTOR_ID: Int = 0x72f0eaae.toInt()
+        const val CONSTRUCTOR_ID: Int = 0x72f0eaae
     }
 }
