+++ tl/src/main/kotlin/com/github/badoualy/telegram/tl/api/TLInputMessagesFilterRoundVoice.kt
@@ -20,6 +20,6 @@ class TLInputMessagesFilterRoundVoice : TLAbsMessagesFilter() {
         return true
     }
     companion object  {
-        const val CONSTRUCTOR_ID: Int = 0x7a7c17a4.toInt()
+        const val CONSTRUCTOR_ID: Int = 0x7a7c17a4
     }
 }
