+++ tl/src/main/kotlin/com/github/badoualy/telegram/tl/api/TLInputNotifyChats.kt
@@ -20,6 +20,6 @@ class TLInputNotifyChats : TLAbsInputNotifyPeer() {
         return true
     }
     companion object  {
-        const val CONSTRUCTOR_ID: Int = 0x4a95e84e.toInt()
+        const val CONSTRUCTOR_ID: Int = 0x4a95e84e
     }
 }
