+++ tl/src/main/kotlin/com/github/badoualy/telegram/tl/api/TLInputReportReasonViolence.kt
@@ -20,6 +20,6 @@ class TLInputReportReasonViolence : TLAbsReportReason() {
         return true
     }
     companion object  {
-        const val CONSTRUCTOR_ID: Int = 0x1e22c78d.toInt()
+        const val CONSTRUCTOR_ID: Int = 0x1e22c78d
     }
 }
