+++ tl/src/main/kotlin/com/github/badoualy/telegram/tl/api/TLInputStickeredMediaPhoto.kt
@@ -47,6 +47,6 @@ class TLInputStickeredMediaPhoto() : TLAbsInputStickeredMedia() {
         return id == other.id
     }
     companion object  {
-        const val CONSTRUCTOR_ID: Int = 0x4a992157.toInt()
+        const val CONSTRUCTOR_ID: Int = 0x4a992157
     }
 }
